requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
anthropic==0.39.0
python-docx==1.1.2
aiofiles==24.1.0
